[JAVA-8150] Update webclient to use retrieve() method instead of deprecated API
